rkt: convert to github.com/spf13/cobra cli library
This patch converts rkt to use the cobra cli library. It doesn't do any
other change. For this reason the help/usage output is different from
the previous one.

Cobra internally uses the github.com/spf13/pflag library (forked from
github.com/ogier/pflag, it's not a dropin replacement for stdlib flag
packages due to an additional method Type() added to the Value
interface). This library handles posix flags and interspersed flags
(flags after non flag arguments).
This means that long flags must start with a double dash (with package
flag also a single dash is accepted).
Until spf13/pflag#20 isn't fixed, an `=` between the long flag and its
value is needed.

Cobra adds the ability to merge the parent command flags (called
persistent flags) to all the child commands. In this way there's no need
to specify the global flags after the executable and before the command
name.

This patch also enables the handling of unambiguous not fully typed
commands (like `rkt l` for `rkt list`).

In rkt there're three special kind of arguments parsing, that needs
special handling of the cobra/pflags default behavior:

fetch: It can handle multiple images with an optional --signature option
after every image. To avoid the --signature being parsed as a normal
flag, interspersed flags must be disabled.

enter: it handles an executable and optional arguments. To avoid these
arguments being parsed as rkt flags, interspersed flags must be
disabled.

run/prepare: it can handle multiple images with image specific
arguments. To do this the image arguments should be inserted between
"--" and "---". This uses a special argument parsing to handle the "--"
that usually makes the flag (and also pflag) libraries stop parsing
flags, eating the "--"  and returning everything after it as arguments
to the relative command. To get the same behavior with the pflags
library, interspersed flags must be disabled and the "--" must be
readded to the command's arguments.