rkt: handle http caching (ETag and Cache-Control max-age)
this patch handles ETag and Cache-Control: max-age server headers.

It saves the Etag, the max-age and the current download time in the store's
remote data.

Before downloading it checks if the current time is before the download time
plus maxage, if so it uses the current store image.
During an ACI download it also sends the saved ETag, if the server answers with
a 304 Not Modified status then the store image is used.

The signature will be downloaded anyway because it's not cached in the store
and it's usually downloaded before the ACI.

This patch also provides db migration code (to version 2) and related tests as
the store's remote struct changed.

Signed-off-by: Simone Gotti <simone.gotti@gmail.com>
Co-Authored-By: Hector Fernandez <hector@elasticbox.com>