rkt: handle multiple rendered images for the same imageID
If an image has dependencies, and they have changed, for the same imageID we
can have multiple different rendered images.

Until now this wasn't handled and rkt and the treestore used the imageID to
reference the rendered image.

This patch, does multiple things:

* the treestore now returns a treeStoreID on store.RenderTreeStore. The
treeStoreID is computed as an hash of the flattened dependency tree image keys.
In this way if some dependencies have changed a new ID will be generated and a
new image rendering will be done.

* rkt has to use this treeStoreID instead of the imageID.

 * for doing this, at prepare phase, the stage1 tree store ID and every app
treeStoreID is saved in the pod.

Additionally in this patch the prepare phase will create a pod with all the
needed data for the run phase without the need to keep the needed images in the
store (excluding the need of the treestore rendered images for overlayfs)
(previously store.GetImageManifest was called in the run phase).

Now `rkt image rm` just removes the image from the blob store.
All the rendered images in the tree store aren't touched by `rkt image rm`.
They will need a future patch for cleaning the unused one (a store gc). This
will probably add a reference count mechanism to the treestore, since just
retrieving the treeStoreIDs from the pods is quite racy (due to pods getting
created between getting the lists and removing the tree store rendered image).

This means that now we can remove any image in the store at any time after the
prepare phase.