rkt: create a new outputDir for every docker2aci call.
Now the output dir provided to docker2aci is always the same (the store
tmpDir).
This will create some confusion or possibly some races if multiple rkt
processes are fetching the same image.

With this patch a new temporary dir under the store tmpDir is created for every
docker2aci call.