*: change appName to name for images
When storing and printing information about ACIs, we were using the
rather incorrect `appName` instead of the more appropriate `name`. This
changes that in both the rkt image list output and in the underlying
store. Since this is one of the columns in the database we use to index
ACI information, it also involves a database schema migration.