rkt/image: set default os/arch for fetching by app name only on top image.
Actually rkt, for fetching an app by name, sets, if not provided by the
image string, os and arch discovery labels to the current machine os/arch.

If an image has a dependency on a noarch image it should define in its
manifest dependency only the app name, version and not os and arch
labels.

This will always break fetching from the store since GetACI won't be
able to find an image with the required labels (because
a noarch image doesn't contain os arch labels in their manifest).

It'll also break some types of meta discovery tags. For example:

```
<meta name="ac-discovery" content="example.com https://storage.example.com/{name}-{version}-{os}-{arch}.{ext}">
<meta name="ac-discovery" content="example.com https://storage.example.com/{name}-{version}.{ext}">
```

will always render the first template. (Ex):
https://storage.example.com/name-version-linux-amd64.aci

This patch will set default os and arch only for the first image (since
it's a good default as an image to be executed cannot be a noarch
image). In future, an option to also disable the adding of default
os/arch for the top image (useful only for fetching but not for run)
will be added.

** IMPORTANT **

This will break image fetching (for some kind of meta discovery tags)
for images that define a dependency without providing explicit os and
arch labels.

for example if meta tags are:

```
<meta name="ac-discovery" content="example.com https://storage.example.com/{name}-{version}-{os}-{arch}.{ext}">
```

This template will fail to render and fetch will fail.

IMHO that images are not completely correct since they don't specify the
required os and arch and it's better to break this now than later.