treestore: new implementation
This patch implements a new treestore to handle different needs:

* Use a key/value db (boltdb) to save store data. The primary key is the
store id. An additional non unique index on the image digest is created
to easy getting all the treestore generated with the same starting image
(needed for ACI, not needed for OCI)

* Move the default treestore dir to datadir/treestore/

* Handle backward compatibility (to handle prepared/exited/garbage pods
with previous version a work with the new implementation) with the
previous treestore implementation and paths.

* The store will in future be extended to also support other image types
than ACI (like OCI images).