rkt: Introduce distribution concept
A distribution represents the method to fetch an image starting from an
input string.

Before this concept was introduced, rkt used the ImageType concept, the
ImageType was mapped to the input image string format and internally
covered multiple concepts like distribution, transport and image type
(hidden since all are now appc ACIs).

The distribution concept will be used as the primary information in the
different layers of rkt (fetching but also for references in a CAS/ref
store).

This patch introduces the distribution concept and three kind of
distributions. It just make the smallest required changes to the fetcher
logic to make them work and pass the functional tests. This is the base
for a future fetchers (a future patch will refactor the fetcher logic to
better fit the distribution concept) and store refactors.

See Documentation/devel/distribution.md for a detailed description.