rkt: implement generic no volume policy
This patch implements a generic no volume policy instead of a custom, per
stage1, behavior based on a docker2aci specific app annotation.

This also avoids stage1 having to read the app image manifest but just uses the
pod manifest as its unique source for app setup.

Since the behavior on unspecified volumes is not defined in the appc/spec but
left to the ACE, the idea is to use a per pod.RuntimeApp annotation to define
the runtimeApp no volume policy.

The annotation is called `coreos.com/rkt/pod/annotations/novolumepolicy` and
can have two possible values:

* `empty`: create an empty, per app, volume.
* `empty-copy`: create an empty, per app, volume and copy image contents.

`empty` is the default if the annotation doesn't exists.

rkt, in the stage0, will check if the annotation is already provided by the pod
manifest and only validate it.
If not provided and the image has been converted using docker2aci then an
annotation with value `empty-copy` will be defined.

This has the additional effect to letting the user define (and override for
docker2aci converted images) the no volume policy for every app (with this
patch only by using --pod-manifest, in future, if needed, a per app command
line flag could be added)

Notes:
Before this patch an user provided pod manifest was only validated without
changing it, now, instead, it's also enhanced.