rkt: implement generic no volume policy
This patch implements a generic no volume policy instead of a custom, per
stage1, behavior based on a docker2aci specific app annotation.

This also avoids stage1 having to read the app image manifest but just uses the
pod manifest as its unique source for app setup.

Since the behavior on unspecified volumes is not defined in the appc/spec but
left to the ACE, the idea is to use a pod.RuntimeApp annotation to define the
no volume policy per runtime app (instead of extending the appc PodManifest
with rkt specific fields).

The annotation is called `coreos.com/rkt/annotations/novolumepolicy` and
can have two possible values:

* `empty`: create an empty, per app, volume.
* `empty-copy`: create an empty, per app, volume and copy image contents.

`empty` is the default if the annotation doesn't exists.

rkt, in the stage0, will check if the annotation is already provided by the pod
manifest and, in such case, only validate it.
If not provided and the image has been converted using docker2aci then an
annotation with value `empty-copy` will be defined.

This has the additional effect to letting the user define (and override for
docker2aci converted images) the no volume policy for every app (with this
patch only by providing a --pod-manifest, in future, if needed, a per app
command line flag could be added)

Note that only pod.RuntimeApp annotations are checked for
`coreos.com/rkt/annotations/novolumepolicy` and not a app manifest
annotation (this, to correctly work, requires #3100).

**Notes**
Before this patch an user provided pod manifest was only validated without
changing it, now, instead, it's also enhanced.